Library: xlink_ready: warmstart basic unless already in direct mode
PreviousÃly theready command would only ping to see if the server is
available, and on success assume that it is ready to use. Trying to
load and run a new basic program on top of an already running basic
programm was not reliable in this case.

Thus if the server can be successfully pinged the first time, we'll
now check that the c64 is in direct mode by peeking into the error
mode flag at $9d. If it is anything but $80 (indicating direct mode)
we'll warm start basic by jumping to $fe66 first (equivalent to
pressing runstop/restore).