Firmware/Library: Major improvement of error handling and propagation
The send and receive primitives now take an explicit timeout argument.
Timeouts are implemented on the device using the hardware timer instead
of doing possibly endless loops and letting the watchdog hard reset the
device. When a timeout occurs in either send or receive, the usb transfer
will be completed nevertheless. The driver then checks whether the transfer
was aborted early by checking the expected vs actual transfer size. It then
returns false and sets xlink_error approriately.

The default timeout value is still one second (matching the previous
watchdog configuration). This means that xlink_send() and xlink_receive()
expect the server to react immediately. If the server needs more time before
servicing the next receive or send, the client may use the newly added
xlink_send/receive_with_timeout() functions, which take an additional
timeout argument (in seconds). The libusb message timeout is always one
second more.

The driver functions send and receive now return a boolean accordingly. The
user visible functions in the xlink library now check these return values
and abort early.

The overall advantage is a) possibility of waiting for a busy server for
more than one second and b) the library can react to hanging transfers in
a sane manner, propagating the error to the user.

Note that the new xlink_send/receive functions allow to set a zero timeout,
meaning no timeout at all. If these transfers hang, the device will be stuck
in an endless loop and need a power-cycle. But this is only important when
using the generic inject/send/receive stuff. It should never happen when
using the high level functions.