Hardware: Added jumper to configure/adapt C64/C128 reset lines
On the C64 this jumper simply needs to be closed to allow resets. No copper
on the front layer around the pins to allow easy bridging.

On the C128, the reset signal at the userport only resets peripherals. The
jumper needs to be left open, and a wire must be run from the top pin to the
real reset line of the C128 to support hardware resets.