Library: Fixed reset/ready behaviour on C128 and for some edge cases.
On the C128 the System Interrupt will stay active quite some time into the
reset procedure, so we need to wait for it to go away after reset before we
can wait for it to come back.

Also if the program we develop happens to disable sysirq but accidently
flips the PA2 line on $dd00 once more (e.g. while configuring vic banking)
we may get a false positive on the initial ping during xlink_ready(). Thus
we now check twice wether we can ping the server at that point.