Servant64: Implemented working serial driver and protocol (linux)
Windows is not supported yet, but that's only a matter of accessing the
serial port. The serial protocol can still be improved (sending 9 bytes
per command all the time, even though most commands would need only one)

The Arduino nano used for the Servant64 hardware has to be equiped with a
10uF Capacitor between reset and ground to disable the automatic reset
feature. Otherwise the arduino would get reset and enter its bootloader
for the first second each time the serial port is opened, which would
require the client to wait for at least a second on each invocation.

Also the current prototype version needs a connection from the left smd pad
of R1 to the Userport Reset line (3rd pad from the right on the top side)
to support hardware resets.