Library/Client: Removed all disk drive commands and functionality
The disk drive commands have never worked reliably or fast enough and
have bloated the code of both the client and library, and have therefore
been removed entirely.

A proper disk backup tool might be created in the future. Such a tool
would need to use proper fastloading and -saving, error handling and
reproduction and support for more than the standard 35 tracks to be truly
useful.

This measure also narrows the scope of the client and library back to
providing only the basic data transfer and control functionality based
on which more specialized tools may be build.