Library: Fixed double initialisation/finalisation on windows
This bug caused the initialisation and finalisation code to run twice on
windows, resulting in a segfault during finalisation (windows reporting
that the "application has stopped working" on exit).

This bug was introduced by adding constructor/destructor attributes to the
respective functions for gcc/clang under Linux/MacOSX and not removing the
windows-only DllMain() function. It seems that on windows both mechanisms
can be used at the same time.